requirements.txt
Flask~=3.0.3
Flask-SQLAlchemy
Flask-Migrate
Flask-Redis
mysqlclient
load_dotenv
apscheduler
python-dotenv~=1.0.1